pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sec-poc"
version = "0.1.0"
description = "Lightweight observable-only PoC for Sovereign Epistemic Commons under No-Meta Governance"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "matplotlib>=3.8",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["sec_poc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

